+++ examples/connect.js
@@ -5,6 +5,6 @@ const client = new Banchojs.BanchoClient(config["irc_user"], config["irc_pass"],
 client.connect().then(() => {
 	console.log("We're online! Now listening for incoming messages.");
 	client.on("PM", (message) => {
-		console.log(message.user+": "+message.message);
+		console.log(message.user.ircUsername+": "+message.message);
 	});
 }).catch(console.error);